refactor: rename models to app and move services to auth
