+++ src/app/auth/services/createUser.go
@@ -1,10 +1,10 @@
-package userService
+package authService
 
 import (
+	"github.com/rodrigoRVSN/beeus-api/src/app/users/entities"
 	database "github.com/rodrigoRVSN/beeus-api/src/infra/config"
 	hash "github.com/rodrigoRVSN/beeus-api/src/infra/helpers"
 	fieldsValidator "github.com/rodrigoRVSN/beeus-api/src/infra/helpers/validator"
-	"github.com/rodrigoRVSN/beeus-api/src/models/users/entities"
 )
 
 func CreateUser(payload *entities.User) error {
