+++ src/app/users/controllers/user-controller.go
@@ -2,8 +2,8 @@ package userController
 
 import (
 	"github.com/gofiber/fiber/v2"
+	"github.com/rodrigoRVSN/beeus-api/src/app/users/entities"
 	database "github.com/rodrigoRVSN/beeus-api/src/infra/config"
-	"github.com/rodrigoRVSN/beeus-api/src/models/users/entities"
 )
 
 func ListUsers(context *fiber.Ctx) error {
