+++ src/infra/router/routes.go
@@ -2,11 +2,13 @@ package routes
 
 import (
 	"github.com/gofiber/fiber/v2"
-	userController "github.com/rodrigoRVSN/beeus-api/src/models/users/controllers"
+	authController "github.com/rodrigoRVSN/beeus-api/src/app/auth/controllers"
+	userController "github.com/rodrigoRVSN/beeus-api/src/app/users/controllers"
 )
 
 func SetupRoutes(app *fiber.App) {
 	app.Get("/users", userController.ListUsers)
-	app.Post("/users/register", userController.CreateUser)
-	app.Post("/users/login", userController.SignInUser)
+
+	app.Post("/auth/register", authController.CreateUser)
+	app.Post("/auth/login", authController.SignInUser)
 }
